game-session-service/pyproject.toml
[project]
name = "game-session-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
 "python-dotenv>=1.1.1",
 "uvicorn>=0.34.3",
]
